Add CKEditor types
