+++ packages/ui/src/common/components/CKEditor/index.ts
@@ -1,1 +1,2 @@
 export * from './CKEditor'
+export { EventInfo } from './types'
