+++ packages/ui/src/common/components/CKEditor/types.ts
@@ -1,0 +1,17 @@
+export declare class Editor {
+  getData: () => string
+  setData: (data: string) => void
+  isReadOnly: boolean
+  model: any
+  editing: any
+  destroy: () => Promise<undefined>
+
+  static create(): Editor
+}
+
+export interface EventInfo {
+  name: string
+  path: any[]
+  return: any
+  source: any
+}
